Moved dump models to spdump.
